GUI: Dashboard, fix type filtering disappearance.
The combobox shouldn't get hide when there are transactions in the wallet, independently on the selected filter. If the user selects a type that it isn't in the wallet, the filter combobox will disappear, leaving the tx list empty and not allowing any other movement.

Github-Pull: #2073
Rebased-From: 6dbe9ad1eadc5fee65a71d68a78aef38f6d5cce4