[BUG] Fix total budget on testnet
the cycle is 144 blocks, not 146.
Use the available GetBlockValue/nBudgetCycleBlocks, instead of
hardcoding the constants.

Note: GetTotalBudget should consider the possibility of changes to the
block value, between consecutive superblocks.
Since the chain already passed the last reduction, and the block value
is fixed at 5, we can leave this as is for now.

Github-Pull: #2101
Rebased-From: bdfaaed277981dce9618b7a335edd744780aebaf