[GUI][BUG] Fix confirmation dialog for dumpwallet/dumpprivkey
Currently the confirmation pop-up is shown only when the RPC command
string is exactly "dumpwallet"/"dumpprivkey" (with no arguments).
Such command strings are not valid, as both RPC functions require one
argument (the filename for "dumpwallet" and the PIVX address for
"dumpprivkey").

So, when the user confirms the dialog, the RPC console returns an error
(as the argument is missing), showing the RPC help.

If the user, instead, enters the correct RPC strings (`dumpwallet
<filename>`/`dumpprivkey <pivxaddress>`), then **the warning dialog is
NOT presented** and the command is executed immediately.

Fix it by doing a pre-validation for the potentially dangerous commands.
Github-Pull: #2122
Rebased-From: 1d63d1aae1e7e305ffbd96d11740abdcb494995b